requirements.txt
streamlit
faiss-cpu
numpy
sentence-transformers
google-generativeai
pypdf
